api-cinefilo-ia/requirements.txt
openai==1.13.3
python-dotenv
langchain==0.1.11
langchain-openai==0.0.8
faiss-cpu==1.8.0
langchain-google-genai
google-generativeai
pillow